[#9] move view code to reusable class
