+++ CustomDynamicFont/Sources/Presentation/Navigator/Navigator.swift
@@ -21,7 +21,7 @@ final class Navigator {
             let selectOSController = SelectOSViewController(viewModel: viewModel, navigator: self)
             return MainNavigationController(rootViewController: selectOSController)
         case .iOS10:
-            return IOS10ViewController(viewModel: UIKitViewModel())
+            return IOS10ViewController()
         case .iOS11Tabbar:
             let tabbarController = UITabBarController()
             tabbarController.setViewControllers(
